A few songs have > 127 chords The sbyte cast wrapped some ChordIDs to the negative, causing an exception when used in an array lookup

+++ PsarcUtil/PsarcConverter.cs
@@ -384,7 +384,7 @@ namespace PsarcUtil
                                     Techniques = ConvertTechniques((NoteMaskFlag)note.NoteMask),
                                     HandFret = (sbyte)note.AnchorFretId,
                                     SlideFret = (sbyte)note.SlideTo,
-                                    ChordID = (sbyte)note.ChordId
+                                    ChordID = note.ChordId
                                 };
 
                                 if ((chordID != -1) && (chordID != songNote.ChordID))
